Add files via upload
The first code file which includes the steps taken for the data gathering, cleaning and management. The gathering took multiple steps via the YouTube API: First gathering the IDs of videos and channels relevant to the query, then the corresponding video transcripts and channel categories based on the IDs. The data was then cleaned and stored in fitting files based on databank normalisation.