Add files via upload
I gathered data from the YouTube API in a multi-step approach in Python with Pandas and BERTopic. First was the creation of a list of IDs of videos and channels relevant to the query and some parts of the relevant metadata. The relevant transcripts of the videos and the channel category were scraped based on the list. Both during the data gathering and after, steps were taken to keep the data clean, such as marking videos with no transcripts or channels with no category. After removing unnecessary data such as timestamps from the transcripts. Then, the data was stored in tables and pickle files according to data normalization